app/requirements.txt
Flask==1.0.2
Flask-Sockets==0.2.1
gunicorn==19.9.0
requests==2.21.0
GoogleAppEngineCloudStorageClient==1.9.22.1